Collapse similar states to reduce the processing effort
Ultimately it doesn't matter for the world if we're looking at `#..#..#` or `#.#...#`, they are identical in terms
of the distribution of the groups, and therefore for the number of arrangements. By collapsing these states into
each other regularly, and only keeping track of how many we collapsed, the processing time needed goes down from
"days" to "seconds".

Multithreading is still left in place, because it will be a good example to steal later, but the initial attempt
of pruning was now completely useless and got removed.