Move the cycle/checksumming into the implementation
Having it here means we don't need to copy the logic twice, which is helpful -- especially since
the tricky part is that the this doesn't just settle, but rather produces a cycle of a certain length.
This allows to skip the long part as soon as we know that length.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch main
# Your branch is ahead of 'origin/main' by 5 commits.
#   (use "git push" to publish your local commits)
#
# Changes to be committed:
#	modified:   src/day14.rs
#